Unify error dialogs
This puts all error messages threw UiOkDialog(), this both gives things
a more unified look, but also make things more portable. If the graphics
system isn't up and running it will drop down to SDL_SimpleMessage and
if that fails the messaged will be printed in the console.

The boot order as been has been optimized to get us a working graphics
system as soon as possible so that more errors can be displayed by the
internal dialogs.